cleanup and remove deprecated components
