+++ package.json
@@ -13,7 +13,6 @@
     "react-router-dom": "^5.1.0",
     "react-scripts": "3.1.1",
     "react-select": "^3.0.8",
-    "react-toastify": "^6.0.8",
     "use-debounce": "^3.0.1"
   },
   "scripts": {
